+++ src/context/PhysicsContext.ts
@@ -1,13 +1,13 @@
 import React from "react";
 import { Engine } from "matter-js";
-import { PhysicsEvent, PhysicsEventType, PhysicsUpdateFunction, Property } from "../types";
+import { PhysicsEvent, PhysicsEventType, PhysicsUpdateFunction, Position, Property, Velocity } from "../types";
 import { VariableProperty } from "../utils";
 
 type PhysicsContextProps = {
   engine: Property<Engine | null>;
   register: (body: Matter.Body, fn: PhysicsUpdateFunction) => () => void;
-  setGravity: (x: number, y: number) => void;
-  setVelocity: (body: Matter.Body, x: number, y: number) => void;
+  setGravity: (velocity: Velocity) => void;
+  setVelocity: (body: Matter.Body, velocity: Velocity) => void;
   addEventListener: (type: PhysicsEventType, fn: (event: PhysicsEvent) => void) => void;
   removeEventListener: (type: PhysicsEventType, fn: (event: PhysicsEvent) => void) => void;
 }